+++ package.json
@@ -12,7 +12,6 @@
   "author": "anomalilies, cubetastic33, davetan2, Glup3, jy761, niradisgust",
   "license": "ISC",
   "dependencies": {
-    "@prisma/client": "^2.19.0",
     "axios": "^0.21.1",
     "common-tags": "^1.8.0",
     "cron": "^1.8.2",
@@ -28,6 +27,7 @@
     "npm": "6.14.8"
   },
   "devDependencies": {
+    "@prisma/client": "^2.19.0",
     "@types/common-tags": "^1.8.0",
     "@types/config": "0.0.38",
     "@types/cron": "^1.7.2",
