+++ prisma/schema.prisma
@@ -10,10 +10,10 @@ generator client {
 model Fishy {
   id                Int        @default(autoincrement()) @id
   userId            String     @unique
-  totalFish         Int?
+  totalFish         Int?       @default(0)
   totalFishGifted   Int        @default(0)
   timesFished       Int?
-  biggestFish       Int?
+  biggestFish       Int?       @default(0)
   totalTrash        Int        @default(0)
   totalCommon       Int        @default(0)
   totalUncommon     Int        @default(0)
